requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
channels==4.2.0
channels-redis==4.2.1
celery==5.4.0
redis==5.0.7
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
propcache==0.3.2
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.4
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
